editing endpoint response for swagger documentation

+++ src/routes/products.js
@@ -16,16 +16,24 @@ router.post('/add', isAdmin, async (req, res) => {
             is_avaliable,
         };
         const data = await pool.query('INSERT INTO products SET ?', [newProduct]);
-        res.status(200).res.json({Message:"Product added", data});
+        res.status(200);
+        res.json({Message:"added product"});
     } catch (error) {
-        res.status(400).res.json({Message:"verify the data and try again"});
+        res.status(400);
+        res.json({Error:"verify the data and try again"});
     }
 });
 
 // all products
 router.get('/', isLoggedIn, async (req, res) => {
     const products = await pool.query('SELECT * FROM products');
-    res.status(200).res.json(products);
+    if(products == "") {
+        res.status(204);
+        res.json({Message:"no products in the database"});
+    } else {
+        res.status(200);
+        res.json(products);
+    }
 });
 
 // one product
@@ -33,9 +41,16 @@ router.get('/:id', isLoggedIn, async (req, res) => {
     try {
         const { id } = req.params;
         const product = await pool.query('SELECT * FROM products WHERE id = ?', [id]);
-        res.status(200).res.json(product);
+        if(product == "") {
+            res.status(404);
+            res.json({Message:"the product you are looking for does not exist"});
+        } else {
+            res.status(200);
+            res.json(product);
+        }
     } catch (error) {
-        res.status(400).res.json({Error:"the product you are looking for does not exist"});
+        res.status(500);
+        res.json({Error:"Internal Server Error"});
     }
 });
 
@@ -43,10 +58,18 @@ router.get('/:id', isLoggedIn, async (req, res) => {
 router.delete('/delete/:id', isAdmin ,async (req, res) => {
     try {
         const { id } = req.params;
-        const product = await pool.query('DELETE FROM products WHERE id = ?', [id]);
-        res.status(200).res.json({Message:"Product deleted", product});
+        const data = await pool.query('SELECT * FROM products WHERE id = ?', [id]);
+        if(data == "") {
+            res.status(404);
+            res.json({Message:"the product you are looking for does not exist"});
+        } else {
+            const product = await pool.query('DELETE FROM products WHERE id = ?', [id]);
+            res.status(200);
+            res.json({Message:"Product deleted"});
+        }
     } catch (error) {
-        res.status(400).res.json({Error:"the product you are looking for does not exist"});
+        res.status(500);
+        res.json({Error:"Internal Server Error"});
     }
 });
 
@@ -54,18 +77,26 @@ router.delete('/delete/:id', isAdmin ,async (req, res) => {
 router.put('/edit/:id', isAdmin, async (req, res) => {
     try {
         const { id } = req.params;
-        const { name, description, price, picture, is_avaliable } = req.body;
-        const newProduct = {
-            name,
-            description,
-            price,
-            picture,
-            is_avaliable,
-        };
-        const product = await pool.query('UPDATE products SET ? WHERE id = ?', [newProduct, id]);
-        res.status(200).res.json({Message:"Product edited", product});
+        const data = await pool.query('SELECT * FROM products WHERE id = ?', [id]);
+        if(data == "") {
+            res.status(404);
+            res.json({Error:"the product you are looking for does not exist or verify the data and try again"});
+        } else {
+            const { name, description, price, picture, is_avaliable } = req.body;
+            const newProduct = {
+                name,
+                description,
+                price,
+                picture,
+                is_avaliable,
+            };
+            const product = await pool.query('UPDATE products SET ? WHERE id = ?', [newProduct, id]);
+            res.status(200);
+            res.json({Message:"Product edited"});
+        }
     } catch (error) {
-        res.status(400).res.json({Error:"the product you are looking for does not exist or verify the data and try again"});
+        res.status(500);
+        res.json({Error:"Internal Server Error"});
     }
 });
